Move client-side-effects to components/, add README.md
